requirements.txt
numpy>=1.19.2
torch>=1.7.0
torchvision>=0.8.1
matplotlib>=3.3.2
scikit-learn>=0.24.0
Pillow>=8.0.0
medmnist>=2.1.0 